youtube_chat_cli_main/requirements.txt
# ============================================================================
# JAEGIS NexusSync Dependencies
# ============================================================================

# ----------------------------------------------------------------------------
# Core CLI & Configuration
# ----------------------------------------------------------------------------
click>=8.0.0
python-dotenv>=1.0.0
requests>=2.25.0
halo>=0.0.31
colorama>=0.4.6

# ----------------------------------------------------------------------------
# Content Processing
# ----------------------------------------------------------------------------
youtube-transcript-api>=0.6.0
beautifulsoup4>=4.9.0
deepmultilingualpunctuation>=1.0.1
pypdf2>=3.0.0
python-docx>=0.8.11
markdown>=3.4.0
html2text>=2020.1.16

# ----------------------------------------------------------------------------
# LangChain & LLM Integration
# ----------------------------------------------------------------------------
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-huggingface>=0.0.1
langgraph>=0.0.20
langchain-core>=0.1.0

# ----------------------------------------------------------------------------
# Google Drive Integration
# ----------------------------------------------------------------------------
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0

# ----------------------------------------------------------------------------
# Vector Stores
# ----------------------------------------------------------------------------
qdrant-client>=1.7.0
chromadb>=0.4.0

# ----------------------------------------------------------------------------
# Embeddings
# ----------------------------------------------------------------------------
sentence-transformers>=2.2.0
openai>=1.0.0

# ----------------------------------------------------------------------------
# Web Search (FREE options)
# ----------------------------------------------------------------------------
tavily-python>=0.3.0
duckduckgo-search>=4.0.0  # FREE fallback, no API key needed

# ----------------------------------------------------------------------------
# OCR & Document Processing
# ----------------------------------------------------------------------------
pytesseract>=0.3.10
pillow>=10.0.0
pdf2image>=1.16.0

# ----------------------------------------------------------------------------
# Background Services
# ----------------------------------------------------------------------------
apscheduler>=3.10.0

# ----------------------------------------------------------------------------
# TTS (Text-to-Speech)
# ----------------------------------------------------------------------------
gtts>=2.5.0
pydub>=0.25.0
edge-tts>=6.1.0

# ----------------------------------------------------------------------------
# Database
# ----------------------------------------------------------------------------
# SQLite is built into Python, no additional package needed

# ----------------------------------------------------------------------------
# MCP Server
# ----------------------------------------------------------------------------
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# ----------------------------------------------------------------------------
# Testing
# ----------------------------------------------------------------------------
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.0.0
pytest-cov>=4.0.0

# ----------------------------------------------------------------------------
# Development Tools
# ----------------------------------------------------------------------------
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# ----------------------------------------------------------------------------
# Utilities
# ----------------------------------------------------------------------------
tqdm>=4.65.0
python-dateutil>=2.8.0

# Force secure urllib3 version to address known CVEs
urllib3>=2.5.0


# ----------------------------------------------------------------------------
# Optional - Headless Browser (Playwright)
# ----------------------------------------------------------------------------
playwright>=1.40.0  # Optional for JS-rendered scraping; install runtime with: `playwright install chromium`
